Revert "Change tokens from i64 to i32 (#68)" (#73)
This reverts commit 2f133e427b3ec06bba6c5b8c9a05544a77ab0987.

+++ ai_edge_torch/generative/examples/gemma/convert_to_tflite.py
@@ -45,9 +45,9 @@ def convert_gemma_to_tflite(
       checkpoint_path, kv_cache_max_len=kv_cache_max_len
   )
   # Tensors used to trace the model graph during conversion.
-  prefill_tokens = torch.full((1, prefill_seq_len), 0, dtype=torch.int)
+  prefill_tokens = torch.full((1, prefill_seq_len), 0, dtype=torch.long)
   prefill_input_pos = torch.arange(0, prefill_seq_len)
-  decode_token = torch.tensor([[0]], dtype=torch.int)
+  decode_token = torch.tensor([[0]], dtype=torch.long)
   decode_input_pos = torch.tensor([0], dtype=torch.int64)
 
   quant_config = quant_recipes.full_linear_int8_dynamic_recipe() if quantize else None
